+++ Libplanet.Extensions.Cocona/Commands/StatsCommand.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Security.Cryptography;
 using Cocona;
 using Libplanet.Blocks;
 using Libplanet.RocksDBStore;
@@ -81,16 +82,17 @@ namespace Libplanet.Extensions.Cocona.Commands
             {
                 BlockDigest blockDigest = store.GetBlockDigest(hash) ??
                     throw Utils.Error($"Failed to load the block {hash}.");
-                BlockHeader block = blockDigest.Header;
+                BlockHeader blockHeader =
+                    blockDigest.GetHeader(_ => HashAlgorithmType.Of<SHA256>());
                 var perceivedTime = store.GetBlockPerceivedTime(hash);
 
                 Console.WriteLine(
-                    $"{block.Index}," +
-                    $"{block.Hash}," +
-                    $"{block.Difficulty}," +
-                    $"{block.Miner}," +
+                    $"{blockHeader.Index}," +
+                    $"{blockHeader.Hash}," +
+                    $"{blockHeader.Difficulty}," +
+                    $"{blockHeader.Miner}," +
                     $"{blockDigest.TxIds.Length}," +
-                    $"{block.Timestamp.ToUnixTimeMilliseconds()}," +
+                    $"{blockHeader.Timestamp.ToUnixTimeMilliseconds()}," +
                     $"{perceivedTime?.ToUnixTimeMilliseconds()}");
             }
         }